Extract logic of finding Flutter icons
We can't easily test setDecorations even with stubs, as the object passed doesn't identify which decoration type it is, so we'll need to test this logic separately.
